+++ ai-logic/firebase-ai-ondevice-interop/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 16.0.0-beta05
+
 - [feature] Added support for on-device structured output generation using `generateObject` (#8395)
 
 # 16.0.0-beta03